+++ php/lib/PathRouter/RegexRouter/RegexException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace pedroac\routing\PathRouter\RegexRouter;
+
+class RegexException extends \RuntimeException
+{
+}