+++ CHANGELOG.md
@@ -7,5 +7,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Path router implementation using regex
 - Path router implementation using an associative array
 - Path routing interfaces